srcs/user_management/app/requirements.txt
anyio==4.4.0
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==43.0.0
Django==5.0.7
django-allauth==64.0.0
django-cors-headers==4.4.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.17.3
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
pillow==10.4.0
priority==2.0.0
psycopg==3.2.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
requests==2.32.3
service-identity==24.1.0
sniffio==1.3.1
sqlparse==0.5.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.2
watchfiles==0.23.0
wsproto==1.2.0
zope.interface==7.0.1
